+++ src/app/personside/dialogpanel/HentOppgaveKnapp.tsx
@@ -17,7 +17,7 @@ import { plukkOppgaver, selectFodselsnummerfraOppgaver } from '../../../redux/re
 import { AppState } from '../../../redux/reducers';
 import { isLoading, RestResource } from '../../../redux/restReducers/restResource';
 import { AsyncDispatch } from '../../../redux/ThunkTypes';
-import { settJobberMedSpørsmålOgSvar } from '../kontrollsporsmal/skjulPåTversAvVinduerUtils';
+import { settJobberMedSpørsmålOgSvar } from '../kontrollsporsmal/cookieUtils';
 
 const HentOppgaveLayout = styled.div`
     text-align: center;
