+++ src/app/personside/kontrollsporsmal/Kontrollsporsmal.tsx
@@ -11,7 +11,7 @@ import IfFeatureToggleOn from '../../../components/featureToggle/IfFeatureToggle
 import { FeatureToggles } from '../../../components/featureToggle/toggleIDs';
 import { AsyncDispatch } from '../../../redux/ThunkTypes';
 import { lukkKontrollSpørsmål } from '../../../redux/kontrollSporsmal/actions';
-import { jobberMedSpørsmålOgSvar, kontrollspørsmålHarBlittLukketForBruker } from './skjulPåTversAvVinduerUtils';
+import { jobberMedSpørsmålOgSvar, kontrollspørsmålHarBlittLukketForBruker } from './cookieUtils';
 import { getFnrFromPerson } from '../../../redux/restReducers/personinformasjon';
 import Undertittel from 'nav-frontend-typografi/lib/undertittel';
